requirements.txt
# Requirements to be installed
Django
djangorestframework
requests